+++ src/deprecated/Makefile.in
@@ -98,6 +98,7 @@ subdir = src/deprecated
 ACLOCAL_M4 = $(top_srcdir)/aclocal.m4
 am__aclocal_m4_deps = $(top_srcdir)/m4/ax_check_compile_flag.m4 \
 	$(top_srcdir)/m4/ax_check_link_flag.m4 \
+	$(top_srcdir)/m4/ax_compare_version.m4 \
 	$(top_srcdir)/m4/ax_cxx_compile_stdcxx.m4 \
 	$(top_srcdir)/m4/libtool.m4 $(top_srcdir)/m4/ltoptions.m4 \
 	$(top_srcdir)/m4/ltsugar.m4 $(top_srcdir)/m4/ltversion.m4 \
@@ -374,6 +375,7 @@ libfftw_LIBS = @libfftw_LIBS@
 libgdal_CPPFLAGS = @libgdal_CPPFLAGS@
 libgdal_LIBS = @libgdal_LIBS@
 libgmt_CPPFLAGS = @libgmt_CPPFLAGS@
+libgmt_LDFLAGS = @libgmt_LDFLAGS@
 libgmt_LIBS = @libgmt_LIBS@
 libmotif_CPPFLAGS = @libmotif_CPPFLAGS@
 libmotif_LIBS = @libmotif_LIBS@
