+++ src/mbtrnav/Makefile.am
@@ -144,9 +144,9 @@ libtnav_la_SOURCES = \
                 terrain-nav/TNavFilter.cpp \
                 terrain-nav/TNavPointMassFilter.cpp \
                 terrain-nav/TNavParticleFilter.cpp \
-				terrain-nav/TNavBankFilter.cpp \
-				terrain-nav/TNavPFLog.cpp \
-				terrain-nav/TerrainMapOctree.cpp \
+                terrain-nav/TNavBankFilter.cpp \
+                terrain-nav/TNavPFLog.cpp \
+                terrain-nav/TerrainMapOctree.cpp \
                 terrain-nav/PositionLog.cpp \
                 terrain-nav/TerrainNavLog.cpp \
                 terrain-nav/mapio.cpp \
@@ -161,9 +161,9 @@ libtnav_la_SOURCES = \
                 terrain-nav/TRNUtils.cpp
 
 libtnav_la_LIBADD = libgeolib.la \
-					libnewmat.la \
-					libqnx.la \
-					-lm -lnetcdf -lpthread
+                libnewmat.la \
+                libqnx.la \
+                -lm -lnetcdf -lpthread
 
 libtrnw_la_LDFLAGS =  -no-undefined -version-info 0:0:0
 
@@ -173,25 +173,25 @@ libtrnw_la_SOURCES = \
                 trnw/trnif_msg.c
 
 libtrnw_la_LIBADD = libtnav.la \
-					libqnx.la \
-					libnewmat.la \
-					-lm -lpthread
+                libqnx.la \
+                libnewmat.la \
+                -lm -lpthread
 
 libnetif_la_LDFLAGS =  -no-undefined -version-info 0:0:0
 
 libnetif_la_SOURCES = \
-				trnw/netif.c \
-				trnw/trnif_proto.c
+                trnw/netif.c \
+                trnw/trnif_proto.c
 
 libnetif_la_LIBADD = libtrnw.la \
-				$(top_builddir)/src/mbtrn/libmframe.la \
-				-lm -lpthread
+                $(top_builddir)/src/mbtrn/libmframe.la \
+                -lm -lpthread
 
 bin_PROGRAMS =   trn-server trn-cli netif-test trnif-test trnifsvr-test trnclient-test #  readlog writelog
 
 
 trn_cli_SOURCES = trnw/trncli_test.c trnw/trn_cli.c
-trn_cli_LDADD = $(top_builddir)/src/mbtrn/libmframe.la libtrnw.la libqnx.la libnewmat.la  libtnav.la  libgeolib.la  
+trn_cli_LDADD = $(top_builddir)/src/mbtrn/libmframe.la libtrnw.la libqnx.la libnewmat.la  libtnav.la  libgeolib.la
 
 trn_server_SOURCES = utils/trn_server.cpp
 trn_server_LDADD = libtnav.la libqnx.la libnewmat.la libtnav.la libgeolib.la
@@ -208,7 +208,7 @@ trnifsvr_test_LDADD = libnetif.la  ${top_builddir}/src/mbtrn/libr7kr.la libtrnw.
 trnclient_test_SOURCES =  utils/trnclient_test.cpp utils/TrnClient.cpp utils/TerrainNavClient.cpp
 trnclient_test_LDADD = libtnav.la
 
-dist_bin_SCRIPTS = 
+dist_bin_SCRIPTS =
 
 CLEANFILES =
 DISTCLEANFILES =
