+++ src/mbtrnav/Makefile.in
@@ -98,6 +98,7 @@ subdir = src/mbtrnav
 ACLOCAL_M4 = $(top_srcdir)/aclocal.m4
 am__aclocal_m4_deps = $(top_srcdir)/m4/ax_check_compile_flag.m4 \
 	$(top_srcdir)/m4/ax_check_link_flag.m4 \
+	$(top_srcdir)/m4/ax_compare_version.m4 \
 	$(top_srcdir)/m4/ax_cxx_compile_stdcxx.m4 \
 	$(top_srcdir)/m4/libtool.m4 $(top_srcdir)/m4/ltoptions.m4 \
 	$(top_srcdir)/m4/ltsugar.m4 $(top_srcdir)/m4/ltversion.m4 \
@@ -621,6 +622,7 @@ libfftw_LIBS = @libfftw_LIBS@
 libgdal_CPPFLAGS = @libgdal_CPPFLAGS@
 libgdal_LIBS = @libgdal_LIBS@
 libgmt_CPPFLAGS = @libgmt_CPPFLAGS@
+libgmt_LDFLAGS = @libgmt_LDFLAGS@
 libgmt_LIBS = @libgmt_LIBS@
 libmotif_CPPFLAGS = @libmotif_CPPFLAGS@
 libmotif_LIBS = @libmotif_LIBS@
@@ -791,9 +793,9 @@ libtnav_la_SOURCES = \
                 terrain-nav/TNavFilter.cpp \
                 terrain-nav/TNavPointMassFilter.cpp \
                 terrain-nav/TNavParticleFilter.cpp \
-				terrain-nav/TNavBankFilter.cpp \
-				terrain-nav/TNavPFLog.cpp \
-				terrain-nav/TerrainMapOctree.cpp \
+                terrain-nav/TNavBankFilter.cpp \
+                terrain-nav/TNavPFLog.cpp \
+                terrain-nav/TerrainMapOctree.cpp \
                 terrain-nav/PositionLog.cpp \
                 terrain-nav/TerrainNavLog.cpp \
                 terrain-nav/mapio.cpp \
@@ -808,9 +810,9 @@ libtnav_la_SOURCES = \
                 terrain-nav/TRNUtils.cpp
 
 libtnav_la_LIBADD = libgeolib.la \
-					libnewmat.la \
-					libqnx.la \
-					-lm -lnetcdf -lpthread
+                libnewmat.la \
+                libqnx.la \
+                -lm -lnetcdf -lpthread
 
 libtrnw_la_LDFLAGS = -no-undefined -version-info 0:0:0
 libtrnw_la_SOURCES = \
@@ -819,21 +821,21 @@ libtrnw_la_SOURCES = \
                 trnw/trnif_msg.c
 
 libtrnw_la_LIBADD = libtnav.la \
-					libqnx.la \
-					libnewmat.la \
-					-lm -lpthread
+                libqnx.la \
+                libnewmat.la \
+                -lm -lpthread
 
 libnetif_la_LDFLAGS = -no-undefined -version-info 0:0:0
 libnetif_la_SOURCES = \
-				trnw/netif.c \
-				trnw/trnif_proto.c
+                trnw/netif.c \
+                trnw/trnif_proto.c
 
 libnetif_la_LIBADD = libtrnw.la \
-				$(top_builddir)/src/mbtrn/libmframe.la \
-				-lm -lpthread
+                $(top_builddir)/src/mbtrn/libmframe.la \
+                -lm -lpthread
 
 trn_cli_SOURCES = trnw/trncli_test.c trnw/trn_cli.c
-trn_cli_LDADD = $(top_builddir)/src/mbtrn/libmframe.la libtrnw.la libqnx.la libnewmat.la  libtnav.la  libgeolib.la  
+trn_cli_LDADD = $(top_builddir)/src/mbtrn/libmframe.la libtrnw.la libqnx.la libnewmat.la  libtnav.la  libgeolib.la
 trn_server_SOURCES = utils/trn_server.cpp
 trn_server_LDADD = libtnav.la libqnx.la libnewmat.la libtnav.la libgeolib.la
 netif_test_SOURCES = utils/netif-test.c
